create a shared py file for common functions; add createDiagram.py unfinished
