+++ python/src/createDiagram.py
@@ -1,0 +1,7 @@
+import codecs
+import argparse
+
+
+
+if __name__ == "__main__":
+    print("Test")
